+++ dvd2xbox/filefactory/d2xfileftp.h
@@ -40,7 +40,7 @@ public:
 
 	virtual int DUMMYFileRead(LPVOID buffer,DWORD dwToRead,DWORD *dwRead){return 0;};
 
-	virtual DWORD FileSeek(long offset, int origin){return 0;};
+	virtual __int64 FileSeek(long offset, int origin){return 0;};
 	virtual int DeleteFile(char* filename);
 	virtual int DeleteDirectory(char* filename);
 	virtual int MoveItem(char* source, char* dest);
