+++ dvd2xbox/filefactory/d2xfileiso.h
@@ -32,7 +32,7 @@ public:
 
 	virtual int DUMMYFileRead(LPVOID buffer,DWORD dwToRead,DWORD *dwRead){return 0;};
 
-	virtual DWORD FileSeek(long offset, int origin){return 0;};
+	virtual __int64 FileSeek(long offset, int origin){return 0;};
 	virtual int DeleteFile(char* filename){return 0;};
 	virtual int DeleteDirectory(char* filename){return 0;};
 	virtual int MoveItem(char* source, char* dest){return 0;};
